requirements.txt
mysql-connector-python
slack-sdk
python-dotenv